Add alternative functionality to upload images
Due to Heroku's emphereal filesystem, it is necessary to store images in the database.
Functionality to upload images has been added to the upload() function and functionality
to download and serve images has been added to the download() function. Both functions will
soon be part of the add and update functions for recipe records.