Add three step recipe adding process
Adding a recipe to the database has been split into three steps. The
first on add-recipe/info which has a form for all data in the Recipe
table. When completed this is added to a session. The second step is
add-recipe/ingredients which has a form for ingredients. Again when
completed this is added to the user's session. The third step is
/add-recipe/submit which lists all data the user has just inputted
and a submit button which commits all data to the database and connects
ingredients to the newly added recipe.